requirements.txt
alembic==1.7.5
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
fastapi-pagination==0.9.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
humanize==3.12.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
Mako==1.1.6
MarkupSafe==2.0.1
Pillow==8.4.0
psycopg2==2.9.2
pycodestyle==2.8.0
pydantic==1.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
reportlab==3.6.3
SQLAlchemy==1.4.27
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
zipp==3.6.0
